feat: shared component between party screens
